apps/nexus/requirements.txt
fastapi
uvicorn
supabase
langgraph
langchain
langchain-groq
pydantic
python-dotenv
asyncio
